ready to launche!

+++ index.js
@@ -7,10 +7,36 @@ const
     bodyParser = require('body-parser'),
     session = require('express-session'),
     flash = require('connect-flash'),
-    multer = require('multer');
+    multer = require('multer'),
+    bcrypt = require('bcrypt'),
+    saltRounds = 10;
 let
     db,
-    Gebruikers;
+    Gebruikers,
+    geliked;
+
+
+// Multer setup
+const opslag = multer.diskStorage({
+    destination: './static/images/profielfotos',
+    filename: function(req, file, cb) {
+        cb(null, file.originalname);
+    }
+});
+let fileFilter = (req, file, cb) => {
+    if (
+        file.mimetype === 'image/png' ||
+        file.mimetype === 'image/jpg' ||
+        file.mimetype === 'image/jpeg'
+    ) {
+        cb(null, true);
+    } else {
+        cb(new Error('Bestands formaat moet : PNG,JPG,JPEG zijn'), false);
+    }
+};
+let upload = multer({ storage: opslag, fileFilter: fileFilter });
+
+
 
 // .env bestand gebruiken
 require('dotenv').config();
@@ -22,9 +48,9 @@ app
     .use(bodyParser.urlencoded({ extended: true }))
     .use(session({
         secret: process.env.SESSION_SECRET,
-        cookie: { maxAge: 60000 },
+        cookie: { maxAge: 90000 },
         resave: false,
-        saveUninitialized: false,
+        saveUninitialized: true,
         secure: true,
     }))
     .use(function(req, res, next) {
@@ -49,22 +75,97 @@ mongo.MongoClient
         console.log(err);
     });
 
-
 // Routing
 app
     .post('/log-in', inloggen)
     .get('/', goHome)
     .get('/registration', registreren)
-    .post('/registrating', gebruikerMaken)
+    .post('/registrating', upload.single('photo'), gebruikerMaken)
     .get('/logout', uitloggen)
     .get('/edit-pass', wachtwoordform)
     .post('/edit', wachtwoordVeranderen)
     .get('/delete', accountVerwijderen)
-    .get('/start', gebruikers)
     .get('/matches', overzichtMatches)
+    .post('/matches', editProfile)
     .get('/findlove', gebruiker1)
-    .post('/:id', like)
-    // .get('/*', error404);
+    .post('/:email', like)
+    .get('/profile', profiel)
+    .get('/readytostart', readyToStart)
+    .get('/:email', disLike)
+    .get('/*', error404);
+
+
+
+function readyToStart(req, res) {
+    if (req.session.loggedIN === true) {
+        Gebruikers
+            .findOne({ email: req.session.user.email })
+            .then(data => {
+                res.render('readytostart');
+            })
+            .catch(err => { console.log(err); });
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+}
+
+// Update profile page
+async function editProfile(req, res, file) {
+    if (req.session.loggedIN === true) {
+        const hashedPassword = await bcrypt.hash(req.body.wachtwoord, saltRounds);
+        const query = { _id: mongo.ObjectId(req.session.user._id) }; // the current user
+        console.log(req.session.user._id);
+        const updatedValues = { // the new data values
+            $set: {
+                'voornaam': req.body.voornaam,
+                'achternaam': req.body.achternaam,
+                'geboortedatum': req.body.geboortedatum,
+                'wachtwoord': hashedPassword,
+                'gender': req.body.gender,
+                'searchSex': req.body.searchSex,
+                'photo': req.file.originalname,
+                'functie': req.body.functie,
+                'bio': req.body.bio
+            }
+        };
+        console.log(updatedValues);
+
+        db.collection('users')
+            .findOneAndUpdate(query, updatedValues)
+
+        .then(data => {
+                console.log('heeft data gevonden');
+                console.log(query);
+                console.log(data);
+                if (data) {
+                    res.render('readytostart');
+                }
+            })
+            .catch(err => {
+                console.log(err);
+            });
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+}
+
+// Profiel
+function profiel(req, res) {
+    if (req.session.loggedIN === true) {
+        Gebruikers
+            .findOne({ email: req.session.user.email })
+            .then(data => {
+                res.render('profile.ejs', { data: data });
+            })
+            .catch(err => { console.log(err); });
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+}
+
 
 // Checkt of er een ingelogde gebruiker is en stuurt aan de hand hiervan de juiste pagina door
 function registreren(req, res) {
@@ -79,65 +180,69 @@ function registreren(req, res) {
 function goHome(req, res) {
     if (req.session.loggedIN === true) {
         req.flash('succes', 'Hoi ' + req.session.user.voornaam);
-        res.redirect('findlove');
+        res.render('readytostart');
     } else {
         res.render('index');
     }
 }
+
 // Maakt de gebruiker aan op post
-
-function gebruikerMaken(req, res) {
-    let data = {
-        'voornaam': req.body.voornaam,
-        'achternaam': req.body.achternaam,
-        'geboortedatum': req.body.geboortedatum,
-        'email': req.body.email,
-        'wachtwoord': req.body.wachtwoord,
-        'gender': req.body.gender,
-        'searchSex': req.body.searchSex,
-        'photo': req.body.photo,
-        'functie': req.body.functie,
-        'bio': req.body.bio
-    };
-
+async function gebruikerMaken(req, res, file) {
+    const hashedPassword = await bcrypt.hash(req.body.wachtwoord, saltRounds);
     // Pusht de data + input naar database (gebruikers = collection('users'))
     Gebruikers
-        .insertOne(data)
-        .then(data => {
-            req.session.user = data;
-            req.session.loggedIN = true;
-            req.flash('succes', 'Hoi ' + req.session.user.voornaam + ', jouw account is met succes aangemaakt');
-            res.render('readytostart');
-            console.log('Gebruiker toegevoegd');
-        })
-        .catch(err => {
-            req.flash('error', err);
+        .insertOne({
+            voornaam: req.body.voornaam,
+            achternaam: req.body.achternaam,
+            geboortedatum: req.body.geboortedatum,
+            email: req.body.email,
+            wachtwoord: hashedPassword,
+            gender: req.body.gender,
+            searchSex: req.body.searchSex,
+            photo: req.file.originalname,
+            functie: req.body.functie,
+            bio: req.body.bio,
+            hasLiked: [],
+            hasNotLiked: []
+        }, done);
+
+    function done(err, data) {
+        if (err) {
+            console.log(err);
+            req.flash('error', 'Oeps er ging iets fout');
             res.render('registration');
-        });
+        } else if (data) {
+            req.flash('succes', 'Account aangemaakt! Log in');
+            res.render('index');
+        }
+    }
 }
+
 // checkt of gebruiker bestaat en logt in door sessie aan te maken met de email als ID (omdat email uniek is)
 // req.Flash('class voor de div', 'het bericht') geeft dat  error/succes bericht door naar de template en daar staat weer code die het omzet naar html
-function inloggen(req, res) {
-    Gebruikers
-        .findOne({ email: req.body.email })
-        .then(data => {
-            if (data.wachtwoord === req.body.wachtwoord) {
-                req.session.user = data;
-                console.log('ingelogd als ' + req.session.user.email);
-                req.flash('succes', 'Hoi ' + req.session.user.voornaam);
-                res.redirect('findlove');
-                req.session.loggedIN = true;
-            } else {
-                req.flash('error', 'Wachtwoord is incorrect');
-                res.render('index');
-                console.log('Wachtwoord is incorrect');
-            }
-        })
-        .catch(err => {
-            console.log(err);
-            req.flash('error', 'Account is niet gevonden');
+async function inloggen(req, res) {
+    const user = await Gebruikers.findOne({ email: req.body.email });
+    if (user == null) {
+        req.flash('error', 'Account is niet gevonden');
+        res.render('index');
+    }
+    try {
+        if (await bcrypt.compare(req.body.wachtwoord, user.wachtwoord)) {
+            req.session.loggedIN = true;
+            req.session.user = user;
+            console.log('Succesvol ingelogd');
+            req.flash('succes', 'Hoi ' + req.session.user.voornaam);
+            res.render('readytostart');
+        } else {
+            req.flash('error', 'Wachtwoord is incorrect');
             res.render('index');
-        });
+            console.log('Wachtwoord is incorrect');
+        }
+    } catch (err) {
+        console.log(err);
+        req.flash('error', 'Er ging iets mis. Probeer opnieuw');
+        res.render('index');
+    }
 }
 
 function wachtwoordform(req, res) {
@@ -176,15 +281,20 @@ function wachtwoordVeranderen(req, res) {
 
 // Deze functie verwijderd het account door eerst te controleren of gebruiker ingelogd is en daarna account te vinden met die email en verwijderd het account en zet de session.loggedIn naar false  + flasht status naar user
 function accountVerwijderen(req, res) {
-    Gebruikers
-        .findOneAndDelete({ email: req.session.user.email })
-        .then(result => {
-            console.log(`Heeft ${result.deletedCount} account verwijderd.`);
-            req.flash('succes', 'Uw account is met succes verwijderd');
-            req.session.user.loggedIN = false;
-            res.render('index');
-        })
-        .catch(err => console.error(`Error: ${err}`));
+    if (req.session.loggedIN === true) {
+        Gebruikers
+            .findOneAndDelete({ email: req.session.user.email })
+            .then(result => {
+                console.log(`Heeft ${result.deletedCount} account verwijderd.`);
+                req.flash('succes', 'Uw account is met succes verwijderd');
+                req.session.loggedIN = false;
+                res.render('index');
+            })
+            .catch(err => console.error(`Error: ${err}`));
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
 }
 // Zet de session.loggedIN naar false = niemand ingelogd. Session destroyen is niet mogelijk, omdat flash sessions nodig heeft
 function uitloggen(req, res) {
@@ -193,63 +303,105 @@ function uitloggen(req, res) {
     res.render('index');
 }
 
-// function pagina gebruiker 1
+// Deze functie toont alle gefilterde gebruikers die nog geliked moeten worden
 function gebruiker1(req, res) {
-    Gebruikers
-        .find({ _id: { $ne: mongo.ObjectId(req.session.user._id) } }).toArray()
-        .then(data => {
-            res.render('detail', { data: data });
-        })
-        .catch(err => { console.log(err); });
-}
-// function pagina gebruiker 1
-function overzichtMatches(req, res) {
-    Gebruikers
-        .find({ _id: { $ne: mongo.ObjectId(req.session.user._id) } }).toArray()
-        .then(data => {
-            res.render('match', { data: data });
-        })
-        .catch(err => { console.log(err); });
-}
-// function db
-function gebruikers(req, res) {
-    Gebruikers
-        .find({ _id: { $ne: mongo.ObjectId(req.session.user._id) } }).toArray()
-        .then(data => {
-            res.render('add', { data: data });
-        })
-        .catch(err => { console.log(err); });
-}
-// Functie liken 
-function like(req) {
-    let id = req.params.id;
-
-    // like toevoegen aan lijst/array hasliked
-    Gebruikers
-        .updateOne({ id: userid }, { $push: { 'hasLiked': id } });
-
-    // like toevoegen aan users liked collection
-    Gebruikers
-        .findOne({ id: id }, addToCollection);
-}
-let matchedStatus;
-
-function addToCollection(err, data, userid) {
-    if (err) {
-        throw err;
+    if (req.session.loggedIN === true) {
+        Gebruikers
+            .find({
+                $and: [
+                    { _id: { $ne: mongo.ObjectId(req.session.user._id) } },
+                    { email: { $nin: req.session.user.hasLiked } },
+                    { email: { $nin: req.session.user.hasNotLiked } },
+                    { gender: req.session.user.searchSex },
+                    { searchSex: req.session.user.gender }
+                ]
+            }).toArray()
+            .then(data => {
+                res.render('detail', { data: data });
+                console.log(data);
+            })
+            .catch(err => {
+                console.log(err);
+                req.flash('error', 'Excuses! er ging iets fout. Probeer het opnieuw');
+                res.render('readytostart');
+            });
     } else {
-
-        if (!data.hasNotliked.includes(userid)) {
-            if (data.hasLiked.includes(userid)) {
-                matchedStatus = true;
-            }
-        }
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
     }
 }
 
-// // Bij een 404
-// function error404(res) {
-//     res.render('404');
-// }
+// functie die kijkt of er matches zijn en deze pusht naar matches om te laten zien op de matchpagina
+function overzichtMatches(req, res) {
+    let matches = [];
+    if (req.session.loggedIN === true) {
+        let gelikedeusers = req.session.user.hasLiked;
+        let huidigemail = req.session.user.email;
+        if (gelikedeusers) {
+            Gebruikers
+                .find({ email: { $in: gelikedeusers } }).toArray()
+                .then(data => {
+                    for (let i = 0; i < data.length; i++) {
+                        if (data[i].hasLiked.includes(huidigemail)) {
+                            matches.push(data[i]);
+                        }
+                    }
+                    res.render('match', { data: matches });
+                })
+                .catch(err => {
+                    console.log(err);
+                    req.flash('error', 'Excuses! er ging iets fout. Probeer het opnieuw');
+                    res.render('readytostart');
+                });
+        } else {
+            req.flash('error', 'U heeft nog geen matches');
+            res.render('match');
+        }
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+
+}
+
+function like(req, res) {
+    if (req.session.loggedIN === true) {
+        let id = req.params.email;
+        console.log(id);
+        Gebruikers.updateOne({ _id: mongo.ObjectId(req.session.user._id) }, {
+            $push: { 'hasLiked': id }
+        });
+        req.session.user.hasLiked.push(id);
+        console.log('liked');
+        res.redirect('/findlove');
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+
+}
+
+function disLike(req, res) {
+    if (req.session.loggedIN === true) {
+        let id = req.params.email;
+        console.log(id);
+        Gebruikers.updateOne({ _id: mongo.ObjectId(req.session.user._id) }, {
+            $push: { 'hasNotLiked': id }
+        });
+        req.session.user.hasNotLiked.push(id);
+        console.log('disliked' + id);
+        res.redirect('/findlove');
+    } else {
+        req.flash('error', 'U moet eerst inloggen');
+        res.render('index');
+    }
+
+}
+
+// Bij een 404
+function error404(req, res) {
+    res.render('404');
+}
+
 // Welke poort het live staat
 app.listen(5000, () => console.log('App is listening on port', port));
